added passport-jwt authentication

+++ controller/api/v1/posts_api.js
@@ -28,27 +28,26 @@ module.exports.destroy = async function (req, res) {
         let post = await Post.findById(req.params.id);
 
         // .id means converting the object id into string
-        // if (post.user == req.user.id) {
-        post.deleteOne();
+        if (post.user == req.user.id) {
+            post.deleteOne();
 
-        await Comment.deleteMany({ post: req.params.id });
+            await Comment.deleteMany({ post: req.params.id });
 
 
 
-        // req.flash('success', 'Post and associated comments deleted!');
+            return res.json(200, {
+                message: "Posts and associated commets deleted"
+            });
 
-        return res.json(200, {
-            message: "Posts and associated commets deleted"
-        });
-
-        // }
-        // else {
-        //     req.flash('error', 'You cannot delete this post!');
-        //     return res.redirect('back');
-        // }
+        }
+        else {
+            return res.json(401, {
+                message: "You cannot delete this post!"
+            });
+        }
     }
     catch (err) {
-        // req.flash('error', err);
+
         return res.json(500, {
             message: "Internal Server Error"
         });
